+++ CHANGELOG.md
@@ -8,9 +8,11 @@ Added
 -
 
 * Add static partitioning in ArmoniK
-* Add ArmoniK configmaps: compute-plane-configma and control-plane-configmap
+* Add ArmoniK configmaps: compute-plane-configmap and control-plane-configmap
 * Add the environment variable `Amqp__PartitionId` in the of ArmoniK polling agent container
 * Add partition metrics exporter service (monitoring)
+* Add a pod to insert partitions data in the database
+* Add partition metrics exporter in Prometheus scraping list
 
 Changed
 -